% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorships.R
\docType{data}
\name{authorships}
\alias{authorships}
\title{Authorships sample}
\format{A data frame with 2641 rows and 5 variables:
\describe{
  \item{title}{The title of an article.}
  \item{authors}{All the authors of the article.}
  \item{value}{A single author of the article - with the title forms an authorship; there can be several authorships per article.}
  \item{genderCoded}{Manually coded gender of an author.
   There are four codes: "female", "male", "noname", "unknown". 
   "Noname" is the code for a case were human coders were not able to find 
   a first name of an author. "Unknown" is the code for a case were 
   the coders found a full name of an author but were not able to verify 
   if she or he is a man or a female.}
  \item{WOSaccessionNumber}{The original ID of an article 
  in WebOfScience database.}
}}
\source{
\url{http://webofknowledge.com/}
}
\usage{
authorships
}
\description{
A dataset containing a simple random sample of authorships 
(unique combination of authors and titles) from WebOfScience records 
of articles of "biographical-items" or "items-about-individual" types
from all fields of study
published from 1945 to 2014.
The sample was drawn in December 2014.
}
\keyword{datasets}
