% Generated by roxygen2 (4.0.2): do not edit by hand
\name{numberOfNames}
\alias{numberOfNames}
\title{Number of names in the database.}
\usage{
numberOfNames()
}
\value{
returns a numeric value
}
\description{
\code{numberOfNames} returns a number of distinct names in database scrapped from genderize.io page.
}
\examples{
\dontrun{

numberOfNames()

}
}

