% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/authorships.R
\docType{data}
\name{authorships}
\alias{authorships}
\title{Authorships sample}
\format{A data frame with 2641 rows and 5 variables:
\describe{
  \item{title}{title of an article}
  \item{authors}{all authors for this article}
  \item{value}{a single author/authorships}
  \item{genderCoded}{manually coded gender of an author}
  ...
}}
\source{
\url{http://webofknowledge.com/}
}
\usage{
authorships
}
\description{
A dataset containing sample of authorships (unique combination of authors
and titles) from WebOfScience records of article of biographical-items
and items-about-individual types.
}
\keyword{datasets}

