% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/authorships.R
\docType{data}
\name{authorships}
\alias{authorships}
\title{Authorships sample}
\format{A data frame with 2641 rows and 5 variables:
\describe{
  \item{title}{title of an article}
  \item{authors}{all authors for this article}
  \item{value}{a single author of the article - with the title forms an authorship}
  \item{genderCoded}{manually coded gender of an author. There are four codes: female, male, noname, unknown. Noname is the code for a case were coders were not be able to find a proper first name of an author. Unknown if the code for a case were coders found a full name of an author but were not be able to verify if she or he is a man of a female.}
  \item{WOSaccessionNumber}{original ID of an article}
}}
\source{
\url{http://webofknowledge.com/}
}
\usage{
authorships
}
\description{
A dataset containing a random sample of authorships (unique combination of authors
and titles) from Web Of Science records of articles from biographical-items
and items-about-individual categories.
}
\keyword{datasets}

