% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/classificationErrors.R
\name{classificatonErrors}
\alias{classificatonErrors}
\title{Calculating classification errors on confusion matrix}
\usage{
classificatonErrors(labels, predictions)
}
\arguments{
\item{labels}{A vector of true labels. Shoud have following
values: c("female", "male", "unknown", "noname"). "noname" stands also for
initials only.}

\item{predictions}{A vector of true labels. Shoud have following
values: c("female", "male", NA).}
}
\value{
A list
}
\description{
\code{classificatonErrors} builds confusion matrix from manually coded
 and predicted gender vectors and returns different
 specific classification errors.
}
\examples{
\dontrun{

x = c("Winston J. Durant, ASHP past president, dies at 84",
"Gold Badge of Honour of the DGAI Prof. Dr. med. Norbert R. Roewer Wuerzburg",
"The contribution of professor Yu.S. Martynov (1921-2008) to Russian neurology",
"JAN BASZKIEWICZ (3 JANUARY 1930 - 27 JANUARY 2011) IN MEMORIAM",
"Maria Sklodowska-Curie")

textPrepare(x)

}
}

