% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classificatonErrors.R
\name{classificatonErrors}
\alias{classificatonErrors}
\title{Calculating classification errors and other prediction indicators}
\usage{
classificatonErrors(labels, predictions)
}
\arguments{
\item{labels}{A vector of true labels. Should have following 
values: c("female", "male", "unknown", "noname"). \code{noname} stands also for 
initials only.}

\item{predictions}{A vector of predicted gender. Should have following 
values: c("female", "male", NA). \code{NA} when it was not possible to predict any gender.}
}
\description{
\code{classificatonErrors} function was misspelled (sorry for that!). 
Now the function has the proper name \code{classificationErrors} (with "i").
Old function name still works, but is deprecated now and will be removed
in future version of the package.
}

