% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/genderizePredict.R
\name{genderizePredict}
\alias{genderizePredict}
\title{Geder prredicting function}
\usage{
genderizePredict(trainedParams, newdata, givenNamesDB)
}
\arguments{
\item{trainedParams}{An output of a \code{genderizeTrain} function with prediction efficiency indicators for different combinations of probability and count values}

\item{newdata}{A character vector for gender prediction}

\item{givenNamesDB}{A dataset with gender data (could be an output of \code{findGivenNames} function)}
}
\value{
A character vector of values: "male", "female" or "unknown"
}
\description{
\code{genderizePredict} predicts gender with the best values of
 'probability' and 'count' parameters.
}

