% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/genderizeR.R
\docType{package}
\name{genderizeR}
\alias{genderizeR}
\alias{genderizeR-package}
\title{An R package for predicting gender from first names}
\description{
The \code{genderizeR} package uses genderize.io API to predict gender from first names extracted from text corpuses. The accuracy of prediction could be control by two parameters: counts of first names in database and probability of gender given the first name.
}
\seealso{
\itemize{
  \item \url{http://genderize.io/} [Homepage of genderize.io API]
  \item \url{https://github.com/kalimu/genderizeR} [R package source code]
  \item \url{http://www.wais.kamil.rzeszow.pl/genderizeR} [R package homepage]
}
}

