% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genderizeR.R
\docType{package}
\name{genderizeR}
\alias{genderizeR}
\alias{genderizeR-package}
\title{Gender Prediction Based on First Names}
\description{
The \code{genderizeR} package uses genderize.io API to predict 
gender from first names extracted from text corpuses. The accuracy 
of prediction could be controlled by two parameters: 
counts of first names in database and probability of gender 
given the first name.
}
\seealso{
\itemize{
  \item \url{http://genderize.io/} [Homepage of genderize.io API]
  \item \url{https://github.com/kalimu/genderizeR} [source code of the latest development version of the R package]
  \item \url{http://www.wais.kamil.rzeszow.pl/genderizeR} [R package homepage]
}
}

