% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/genderizeTrain.R
\name{genderizeTrain}
\alias{genderizeTrain}
\title{Training genderize function}
\usage{
genderizeTrain(x, y, givenNamesDB, probs, counts)
}
\arguments{
\item{x}{A text vector that we want to genderize}

\item{y}{A text vector of true gender labels for x vector}

\item{givenNamesDB}{A dataset with gender data (could be an output of \code{findGivenNames} function)}

\item{probs}{A numeric vector of different probability values.
Used to subseting a givenNamesDB dataset}

\item{counts}{A numeric vector of different count values.
Used to subseting a givenNamesDB dataset}
}
\value{
A data frame with all combination of parameters and computed
sets of prediction indicators for each combination:
  \item{errorCoded}{classification error for predictet & unpredicted gender}
  \item{errorCodedWithoutNA}{for predicted gender only}
  \item{naCoded}{of manually coded gender only }
  \item{errorGenderBias}{net gender bias error}
}
\description{
\code{genderizeTrain} predicts gender and checks different combination
of 'probability' and 'count' paramters.
}
\examples{
\dontrun{


}
}

