% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/givenNamesDB_titles.R
\docType{data}
\name{givenNamesDB_titles}
\alias{givenNamesDB_titles}
\title{Gender data for titles sample}
\format{A data.table object with 872 rows and 4 variables:
\describe{
  \item{name}{first name}
  \item{gender}{predicted gender}
  \item{probability}{how many persons in with this first name
  has the predicted gender}
  \item{count}{how many persons in the genderize.io database
  had that first name}
}}
\source{
\url{http://genderize.io/}
}
\usage{
givenNamesDB_titles
}
\description{
A dataset with a gender data from genderize.io for the sample
of \pkg{titles} in this package. This is the output
of \code{findGivenNames}
function output is dated on January 12, 2015.
}
\keyword{datasets}

