% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberOfNames.R
\name{numberOfNames}
\alias{numberOfNames}
\title{Number of names in the database.}
\usage{
numberOfNames()
}
\value{
returns a numeric value
}
\description{
\code{numberOfNames} returns a number of distinct names in the genderize.io database scrapped from genderize.io page.
}
\examples{
\donttest{

numberOfNames()

}




}
