% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textPrepare.R
\name{textPrepare}
\alias{textPrepare}
\title{Preparing text vector for gender prediction}
\usage{
textPrepare(x, textPrepMessages = FALSE)
}
\arguments{
\item{x}{A vector of character strings.}

\item{textPrepMessages}{If TRUE verbose output of the prepairing process 
is shown on the console.}
}
\value{
A vector of unique terms with at least two characters.
}
\description{
\code{textPrepare} Takes a text vector and converts it into a vector 
of unique terms. This function is used by default 
by the \code{findGivenNames} function
as text pre-processor before sending a query to the API.
}
\examples{
\dontrun{

x = c("Winston J. Durant, ASHP past president, dies at 84", 
      "Gold Badge of Honour of the DGAI Prof. Dr. med. Norbert R. Roewer Wuerzburg",
      "The contribution of professor Yu.S. Martynov (1921-2008) to Russian neurology", 
      "JAN BASZKIEWICZ (3 JANUARY 1930 - 27 JANUARY 2011) IN MEMORIAM", 
      "Maria Sklodowska-Curie")

head(textPrepare(x))
# [1] "ashp"         "at"           "badge"        "baszkiewicz" 
# [5] "contribution" "curie

}

}

