% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/titles.R
\docType{data}
\name{titles}
\alias{titles}
\title{Titles sample}
\format{A data frame with 2641 rows and 2 variables:
\describe{
  \item{title}{title of an article}
  \item{genderCoded}{manually coded gender of an author}
}}
\source{
\url{http://webofknowledge.com/}
}
\usage{
titles
}
\description{
A dataset containing a random sample of article titles
from WebOfScience records of
biographical-items types.
}
\keyword{datasets}

